Move build_device_info to utils
